Rename MFE CAS22 classes to follow IFE conventions and preserve order
